# driving/api-rest/contracts/examples-openapi-v3.yaml
openapi: 3.0.1
info:
  title: Product API
  description: Product API definition
  version: 1.0.0
...
paths:
  /products:
    get:
      tags:
        - products
      summary: Method to retrieve a collection of products
      description: This method retrieves a collection of products
      operationId: getProductCollection
      parameters:
        - $ref: '#/components/parameters/firstPage'
        - $ref: '#/components/parameters/pageSize'
        - $ref: '#/components/parameters/sort'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPageResponse'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
      security:
        - adfs: [ ]
    post:
      tags:
        - products
      summary: Method to add a new product
      description: This method create a new product
      operationId: createProduct
      requestBody:
        description: productRequest
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
      security:
        - adfs: [ ]
  /products/{productId}:
    get:
      tags:
        - products
      summary: Method to retrieve one product
      description: This method retrieves a product specific by identifier
      operationId: getProductById
      parameters:
        - name: productId
          in: path
          description: Product Id
          required: true
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/locale'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
      security:
        - adfs: [ ]
  /products/product-bar-code/{codeBar}:
    get:
      tags:
        - products
      summary: Method to retrieve one product
      description: This method retrieves a product specific by bar code
      operationId: getCodeBarProduct
      parameters:
        - name: codeBar
          in: path
          description: codeBar
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/locale'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
      security:
        - adfs: [ ]
  /stores/{storeId}:
    get:
      tags:
        - stores
      summary: Method to retrieve one store
      description: This method retrieves a store defined by a specific id
      operationId: getStore
      parameters:
        - name: storeId
          in: path
          description: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreResponse'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
      security:
        - adfs: [ ]
    put:
      tags:
        - stores
      summary: Method to update a store
      description: This method update a store defined by its id
      operationId: updateStore
      parameters:
        - name: storeId
          in: path
          description: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: storeRequest
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreRequest'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreResponse'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
      security:
        - adfs: [ ]
      x-codegen-request-body-name: storeRequest
    delete:
      tags:
        - stores
      summary: Method to delete a store
      description: This method delete a store defined by its id
      operationId: deleteStore
      parameters:
        - name: storeId
          in: path
          description: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        202:
          description: No content
          content: { }
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
      security:
        - adfs: [ ]
  /stores/search:
    post:
      tags:
        - stores
      summary: Method to search/filter stores
      description: This method retrieves stores filtered by criteria with pagination and sorting
      operationId: searchStores
      parameters:
        - $ref: '#/components/parameters/firstPage'
        - $ref: '#/components/parameters/pageSize'
        - $ref: '#/components/parameters/sort'
      requestBody:
        description: Filter criteria to search stores
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreFilterRequest'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorePageResponse'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
      security:
        - adfs: [ ]
  /stores:
    post:
      tags:
        - stores
      summary: Method to create a store
      description: This method creates a store
      operationId: createStore
      requestBody:
        description: storeRequest
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreRequest'
        required: true
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreResponse'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
      security:
        - adfs: [ ]
      x-codegen-request-body-name: storeRequest
components:
  schemas:
    ProductRequest:
      title: productRequest
      type: object
      properties:
        name:
          type: string
          required: true
        code:
          pattern: ^[0-9]{8}$
          type: string
          required: true
        category:
          type: integer
          format: int64
        state:
          $ref : '#/components/schemas/ProductTypeRequestEnum'
    StoreRequest:
      title: StoreRequest
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          pattern: ^.{0,100}$
          type: string
          required: true
        code:
          pattern: ^[0-9]{6}$
          type: string
          required: true
        country:
          type: integer
          format: int64
          required: true
    StorePageResponse:
      title: StorePageResponse
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/StoreResponse'
        pagination:
          $ref: '#/components/schemas/PaginationResponse'
    StoreFilterRequest:
      type: object
      properties:
        name:
          type: string
          example: "La esprilla"
        code:
          type: string
          example: "003117"
        country:
          type: string
          example: "España"
        product:
          type: string
          example: "003117"
        category:
          type: integer
          format: int64
          example: 1
    StoreResponse:
      title: StoreResponse
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        code:
          type: string
    ProductPageResponse:
      title: ProductPageResponse
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
        pagination:
          $ref: '#/components/schemas/PaginationResponse'
    ProductResponse:
      title: ProductResponse
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        code:
          type: string
        state:
          $ref: '#/components/schemas/ProductTypeRequestEnum'
    ProductTypeRequestEnum:
      type: string
      required: true
      enum:
        - ACTIVE
        - INACTIVE
    ErrorResource:
      title: ErrorResource
      type: object
      properties:
        code:
          type: string
        description:
          type: string
        details:
          type: array
          items:
            type: string
...
    PaginationResponse:
      type: object
      properties:
        requestedPage:
          type: integer
          format: int32
        requestedSize:
          type: integer
          format: int32
        retrievedResults:
          type: integer
          format: int32
        totalResults:
          type: integer
          format: int64
        nextPage:
          type: string
        previousPage:
          type: string
...
  parameters:
    locale:
      name: locale
      in: query
      description: Language locale
      required: false
      schema:
        type: string
        pattern: '^[a-z]{2}([_-][A-Z]{2})?$'
        default: 'es-ES'
        example: 'es-ES'
    firstPage:
      name: firstPage
      in: query
      description: Requested page (first page is page 1). If not specified 1 is
        used. [Cannot be negative or zero]
      schema:
        type: integer
        format: int32
        default: 1
    pageSize:
      name: pageSize
      in: query
      description: Number of retrieved elements per page. if not specified 10 is
        used. [Cannot be less than 1]
      schema:
        type: integer
        format: int32
        default: 10
    sort:
      name: sort
      in: query
      description: 'Sort of the query. [sort=+id] will order asc by id, or [sort=+id] desc by id). Default unsorted.'
      schema:
        type: string
  responses:
...
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResource'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResource'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResource'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResource'
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResource'
  securitySchemes:
    adfs:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: ""
          scopes: { }
    aad:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: ""
          scopes: { }